docs: note session hook and false negatives component
