+++ src/hooks/useSessionManager.test.js
@@ -1,12 +1,11 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { render, screen, fireEvent } from '@testing-library/react';
-import React from 'react';
-import App from './App.jsx';
+import { renderHook, act } from '@testing-library/react';
+import useSessionManager from './useSessionManager.js';
 
 // In-memory stub store to simulate IndexedDB-backed persistence
 const memoryStore = { last: null };
 
-vi.mock('./utils/db', () => {
+vi.mock('../utils/db', () => {
   return {
     putLastSession: vi.fn(async (session) => {
       memoryStore.last = session;
@@ -20,78 +19,98 @@ vi.mock('./utils/db', () => {
   };
 });
 
-describe('App persistence flow', () => {
+describe('useSessionManager hook', () => {
+  const baseState = () => ({
+    summaryData: [],
+    detailsData: [],
+    clusterParams: {},
+    dateFilter: { start: null, end: null },
+    rangeA: { start: null, end: null },
+    rangeB: { start: null, end: null },
+    fnPreset: 'balanced',
+    setClusterParams: vi.fn(),
+    setDateFilter: vi.fn(),
+    setRangeA: vi.fn(),
+    setRangeB: vi.fn(),
+    setSummaryData: vi.fn(),
+    setDetailsData: vi.fn(),
+  });
+
   beforeEach(() => {
-    // Ensure a clean store and timers per test
     memoryStore.last = null;
     vi.useFakeTimers();
-    // Clear persistEnabled between tests
     try {
       window.localStorage.removeItem('persistEnabled');
     } catch {
       // ignore storage errors
     }
   });
+
   afterEach(() => {
     vi.useRealTimers();
   });
 
   it('debounces auto-save when enabled and supports save/load/clear controls', async () => {
-    render(<App />);
+    const { result } = renderHook(() => useSessionManager(baseState()));
 
-    const remember = screen.getByLabelText(/remember data locally/i);
-    const saveNow = screen.getByRole('button', { name: /save session now/i });
-    const loadSaved = screen.getByRole('button', {
-      name: /load saved session/i,
-    });
-    const clearSaved = screen.getByRole('button', {
-      name: /clear saved session/i,
-    });
-
-    // Initially disabled until opt-in
-    expect(saveNow).toBeDisabled();
-
-    // Enable persistence; localStorage flag should be set and a debounced save should occur
-    fireEvent.click(remember);
+    act(() => result.current.setPersistEnabled(true));
     vi.advanceTimersByTime(600);
 
     const { putLastSession, getLastSession, clearLastSession } = await import(
-      './utils/db'
+      '../utils/db'
     );
     expect(putLastSession).toHaveBeenCalledTimes(1);
     expect(memoryStore.last).toBeTruthy();
-    expect(memoryStore.last).toHaveProperty('clusterParams');
 
-    // Manual save should call put again
-    fireEvent.click(saveNow);
+    await act(async () => {
+      await result.current.handleSaveNow();
+    });
     expect(putLastSession).toHaveBeenCalledTimes(2);
 
-    // Load should fetch the saved session
-    fireEvent.click(loadSaved);
+    await act(async () => {
+      await result.current.handleLoadSaved();
+    });
     expect(getLastSession).toHaveBeenCalledTimes(1);
 
-    // Clear should remove the saved session
-    fireEvent.click(clearSaved);
+    await act(async () => {
+      await result.current.handleClearSaved();
+    });
     expect(clearLastSession).toHaveBeenCalledTimes(1);
     expect(memoryStore.last).toBeNull();
   });
 
   it('skips invalid duration strings when loading a saved session', async () => {
-    const { buildSession } = await import('./utils/session');
-    const stats = await import('./utils/stats');
+    const stats = await import('../utils/stats');
     const spy = vi.spyOn(stats, 'parseDuration');
 
+    const { buildSession } = await import('../utils/session');
     memoryStore.last = buildSession({
       summaryData: [
-        { Date: '2021-01-01', 'Total Time': '1:00:00', AHI: '1', 'Median EPAP': '5' },
-        { Date: '2021-01-02', 'Total Time': 'bad', AHI: '2', 'Median EPAP': '6' },
+        {
+          Date: '2021-01-01',
+          'Total Time': '1:00:00',
+          AHI: '1',
+          'Median EPAP': '5',
+        },
+        {
+          Date: '2021-01-02',
+          'Total Time': 'bad',
+          AHI: '2',
+          'Median EPAP': '6',
+        },
       ],
       detailsData: [],
     });
 
-    render(<App />);
-    const loadSaved = screen.getByRole('button', { name: /load saved session/i });
-    fireEvent.click(loadSaved);
+    const state = baseState();
+    state.setSummaryData = (rows) => {
+      rows.forEach((r) => stats.parseDuration(r['Total Time']));
+    };
+    const { result } = renderHook(() => useSessionManager(state));
+
+    await act(async () => {
+      await result.current.handleLoadSaved();
+    });
 
     await vi.waitFor(() => expect(spy).toHaveBeenCalled());
     expect(spy.mock.calls.some((c) => c[0] === 'bad')).toBe(true);
